Fix: GPIO_MULTI

+++ gpio/gpio.c
@@ -49,11 +49,11 @@ int32_t gpio_genericInit(struct gpio *g) {
 #ifdef CONFIG_GPIO_MULTI
 struct gpio *gpio_init(uint32_t index);
 int32_t gpio_deinit(struct gpio *gpio);
-struct gpio_pin *gpioPin_init(struct gpio *gpio, uint8_t pin, enum gpio_direction dir, enum gpio_setting setting);
+struct gpio_pin *gpioPin_init(struct gpio *gpio, uint32_t pin, enum gpio_direction dir, enum gpio_setting setting);
 int32_t gpioPin_deinit(struct gpio_pin *pin);
 int32_t gpioPin_enableInterrupt(struct gpio_pin *pin);
 int32_t gpioPin_disableInterrupt(struct gpio_pin *pin);
-int32_t gpioPin_setCallback(struct gpio_pin *pin, bool (*callback)(struct gpio_pin *pin, uint8_t pinID, void *data), void *data, enum gpio_interrupt inter);
+int32_t gpioPin_setCallback(struct gpio_pin *pin, bool (*callback)(struct gpio_pin *pin, uint32_t pinID, void *data), void *data, enum gpio_interrupt inter);
 int32_t gpioPin_setDirection(struct gpio_pin *pin, enum gpio_direction dir);
 int32_t gpioPin_setSetting(struct gpio_pin *pin, enum gpio_setting setting);
 int32_t gpioPin_SchmittTrigger(struct gpio_pin *pin, bool schmitt);
